refactor: extract getVerificationCodeController into its own file
